Add option to activate app on launch
Before this change, a new window is always requested instead of activating
the application, which is not always desired.

Introduce a new setting to align the behavior with the app launcher
implemented in GNOME Shell: open new window only if requested by the user
with Ctrl+Enter, Ctrl+click or middle click, otherwise activate the app.

GNOME Shell implementation:
https://gitlab.gnome.org/GNOME/gnome-shell/-/blob/2266a4539c65310f46c3f40f3a72f38becb19b93/js/ui/appDisplay.js?page=4#L3117-3135